+++ config.js
@@ -1,0 +1,9 @@
+const HIT = {
+	head: 30,
+	body: 25,
+	foot: 20,
+};
+
+const ATTACK = ["head", "body", "foot"];
+
+export { HIT, ATTACK };
